The new macros FHTAGN_VARIANT_CHECK() and FHTAGN_VARIANT_SAFE_GET() should be used instead of using fhtagn::variant::check() and fhtagn::variant::safe_get() respectively. Both functions are fine as they are, but must be called with const variables/references to always produce the desired result - something the macros enforce.
The operator[] functions now also don't throw on errors, but always return the
invalid value. The upside is that this enables the above changes to always work -
the downside is that errors are very unspecific, i.e. you don't know which part
in a chain of [] calls (i.e. x["foo"][0]["bar"]) failed. It's only when using
the resulting variable that errors may occur.