Fixes #27:
boost::thread constructor /copies/ whatever you pass to it. In case of a
function pointer, that's irrelevant. In case of boost::bind, that messes up when
you pass a reference to an object to boost::bind - via boost::ref, for example.
In the bound function, the 'this' pointer points to some unintialized value...
Passing a pointer fixes things.