This is an odd one. It mostly fixes the syntax of the unit test code to also work on MSVC (see #44).
The original syntax worked fine on GCC, not because it did what was expected,
but because the whole code was optimized away.

This patch fixes the test code syntax, and also forbids default construction
of the mandatory type. There's not much point in that, unless one also gives
the type full value semtantics - and then the point would be even more
dubious.